Working Beta - Initial Checkin
Description: Most of the changes are due to int vs int64. Welcome issue decommed and backend labels query tweaked.

Notes:
Note: Updated Gob in Frontend to inlcude RepoID.

+++ pipeline/backend/database.go
@@ -17,7 +17,7 @@ type MemSQL struct {
 type Integration struct {
 	RepoID         int64
 	AppID          int
-	InstallationID int
+	InstallationID int64
 }
 
 func (m *MemSQL) Open() {
